\name{friedland-package}
\alias{friedland-package}
\alias{friedland}
\docType{package}
\title{
\packageTitle{friedland}
}
\description{
\packageDescription{friedland}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{friedland}
\packageIndices{friedland}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{friedland}

Maintainer: \packageMaintainer{friedland}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{ \code{\link{adf}} \code{\link{UnitRoot}} \code{\link{Asymmetry}} \code{\link{NhemShem}} \code{\link{MacKinnon}} \code{\link{aic}} \code{\link{gatelyPriceAsymm}} }
\examples{
~~ simple examples of the most important functions ~~
}
